添加秘书服务

+++ sciman-main/src/main/java/com/sciman/dao/SecretaryMapper.java
@@ -1,6 +1,7 @@
 package com.sciman.dao;
 
 import com.sciman.pojo.Secretary;
+import com.sciman.vo.person.SecretaryView;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
 
@@ -12,7 +13,7 @@ public interface SecretaryMapper {
     List<Secretary> getAllSecretary();
 
     @Select("select * from secretary where employee_id = #{employeeId}")
-    Secretary getSecretaryById(Integer employeeId);
+    Secretary getSecretaryById(Long employeeId);
 
 
 }
